去除后端编译告警
# message auto-generated for no-merge-commit merge:
merge dev-remove-warn into master

去除后端编译告警

Created-by: insightDeveloper

Match-id-a9bc6bf98b28fe9f2c6c50ac82f4f66f1ee58c3f

+++ server/src/modules/advisor/process/AffinityAPIAdvisor.cpp
@@ -138,7 +138,7 @@ bool AffinityAPIAdvisor::CheckApiSeqWithRule(const std::vector<std::string> &rul
         return false;  // 真实数据长度 < 预期数据长度，无法匹配
     }
 
-    for (int i = 1; i < rule.size(); ++i) { // 上文已匹配索引为0的数据
+    for (size_t i = 1; i < rule.size(); ++i) { // 上文已匹配索引为0的数据
         std::string tmp = dataList[index + i].name;
         std::vector<std::string> list = StringUtil::Split(rule[i], "\\|");
         if (std::find(list.begin(), list.end(), tmp) == list.end()) { // 不完全匹配，则跳过
